fix - unload-reload same scene during disconnect period
This fixes the issue where a client could be disconnected. the server unload and reloads the same scene while the client is disconnected, and then the client reconnects using the NetworkTableState prior to the same scene being unloaded and reloaded.  The issue was that in the GetAndAddNewlyLoadedSceneByName we needed to exclude any scenes that are in the NetworkTableState as they are the previously loaded scenes and GetAndAddNewlyLoadedSceneByName is only invoked if a scene has to be loaded and a scene only has to be reloaded if the NetworkSceneHandle doesn't already exist in the NetworkTableState.