update
Added additional debug info when a scene load fails.
Added the case for the same scene to be unloaded and reloaded while a client is disconnected and then the client connects.  This added test verifies that the scenes in the NetworkSceneTable state are not taken into consideration when trying to find a newly loaded scene.