fix: NetworkSceneManager does not synchronize despawned in-scene placed NetworkObjects [MTT-2924] (#1898)
* fix

MTT-2924
This is the first pass at solving the issue where late joining clients are not synchronized properly with despawned in-scene NetworkObjects.

* Update CHANGELOG.md

* Update CHANGELOG.md

* fix

This allows for respawning of in-scene placed NetworkObjects without having to register them with the NetworkManager.

* Update CHANGELOG.md

* style

* Update CHANGELOG.md

* style

whitespace fixes.

* fix

Passing the entire SceneObject structure as opposed to adding more parameters to the NetworkSpawnManager.HasPrefab method in order to be able to extract the NetworkSceneHandle and include the in-scene placed NetworkObjects as part of the "prefabs to search for".

* update

This is some work towards integration tests where clients might be able to actually load scenes as opposed to fake loading them

* fix

Fixed some minor issues with tests.
Updated HiddenVariableTests as it needed some timing related adjustments.

* Update NetworkSceneManagerEventDataPoolTest.cs

* style

* style

* test - integration test clients now load scenes

This update allows for clients to load scenes and synchronize in-scene placed NetworkObjects during integration testing.

* test - fix

Now that clients are loading scenes, needed to make some minor adjustments.

* test - fix

Realized I don't need to simulate client loading delay unless the client is blocked from loading.
Also had to adjust whether the metrics tests waited for the client or not when loading.

* fix

Fixes an issue that can occur where this becomes enabled in test runner and will throw an exception if target is null.

* fix - NetworkSceneManager.GetSceneRelativeInSceneNetworkObject

GetSceneRelativeInSceneNetworkObject needed to get the sceneHandle before checking for it existing in the ScenePlacedObjects.

* fix - AddDespawnedInSceneNetworkObjects

AddDespawnedInSceneNetworkObjects needed to only return despawned NetworkObjects belonging to the NetworkManager.

* test - fix - Integration test client scene loading

This fixes some of the last edge case scenarios such as:
- Not loading scenes for clients when a non-NetcodeIntegrationTest based integration test is running since this does require some additional management and cleanup.
- Assuring that in-scene NetworkObjects have the right NetworkManager assigned when loaded
- Added a check to assure that it didn't time out while waiting for all clients to connect
- Increased the NetcodeIntegrationTestHelpers default timeout to 2s vs 1s (clients loading scenes can slow down when things happen)
- NetcodeIntegrationTests now unload all scenes loaded (but the test runner scene) during teardown and onetimeteardown.

* test update

Just adding timeout checks in both NetworkVariableTests and TimeMultiInstanceTest

* test assets

Updating some of the assets used for integration testing now that clients are loading scenes too.

* test fix

Changing the scenes being loaded to avoid integration test specific SDK architecture related issues.

* test fix

An early shutdown was not needed here

* test fix

Disabling the audio listener in the EmptyScene to reduce console spam during integration tests.

* test fix

Fixing edge case timing issue with MessageOrdering.

* test InScenePlacedNetworkObjects

This is the integration test for this PR.
It tests late joining client synchronization with despawned NetworkObjects and tests that in-scene placed NetworkObjects do not need to be registered with the NetworkManager to be spawned.

* fix

Don't throw an exception when calling GetObservers or IsNetworkVisibleTo when the NetworkObject is not spawned.

* fix

fixing typo/mistake for the scene to be unloaded.

* style

* Update ObjectNameIdentifier.cs

* style

* style

updating comments and removing LFs and some commented out code.

* update and style

updating comments and minor final tweaks.

* update

minor changes suggested by Jeffrey R.

* update

removing scene management specific code from places where scene management is disabled.

* style

updated some comments

* fix

Fixing an issue where CheckValueOnClient was crashing.  Integration test timing did change slightly now that clients load scenes, so it was possible that the key would not exist by the time it was being checked.

* fix  consoles

reducing the number of clients from 9 to 4 in some of the NetworkSceneManager related tests as I think it impacted performance enough to cause clients to take longer than the default timeout of 2 seconds.

* fix

Hopefully this will fix the issue with Xbox Scarlett

* test - fix

Removing m_CanCheckLoadUnloadEvents as that was a sanity check while I was testing clients loading scenes.  This could possibly fix an edge case scenario issue where a message might get ignored.

* update

removing NetworkSceneHandle from NetworkObject as it was not needed.

* Update CHANGELOG.md

* update

Removing wait for clients to unload scenes and unloading them via SceneManager as the test has already completed and this is more clean up than anything else.

* test update

Removed the crazy code in ObjectNameIdentifier for adding tag information and simplified it by storing the original name and using that each time it needs to regenerate the tag information.
Moved where the NetworkManager name and tag information is applied to get the right client id.
Added additional debug information.

* style

* test

Testing with verbose mode to see why this passes on Windows 10 in a stand alone build and in the editor but fails on Yamato desktops.

* test

adding more debug info.

* test fix

Looks like the scene isn't really loaded when it says it is loaded.
Moving when the NetworkSceneManager's callback is invoked and adding delay before invoking.

* test fix

Making some additional modifications to see if it was when the NetworkSceneManager scene Load/Unload event action is invoked and reducing the number of clients to 1.

* Test Fix

Minor tweaks and merging develop into this branch.

* test fix

Found a bug in some recent changes that would cause a crash under specific scenarios.

* fix

* test and style

Went back through some of the changed files and removed some code no longer needed, extended some of the verbose debug info for any future debugging purposes, and updated some of the comments.

* style

removing commented out code.