test
This verifies that late joining clients get synchronized to the current NetworkAnimator state.
This also fixes the issue where stopping a client during an integration test always meant that you wanted to delete it.  Now, you have the option to not destroy in the event you want to stop a client, do some stuff, and then reconnect the client.