fix
This reverts Triggers back to being set by an RPC (better to separate for now).
This also fixes minor issues with a player joining right in the middle of a transition.
This also fixes issue where trigger paths were not being set when transitioning (still could use more work).
This fixes the issue with the NetworkAnimator failing on the late joining player test where the late joining player could join while an Animator is in the middle of a transition.