fix: NetworkAnimator does not update animator's parameters unless the animator's layer changes [MTT-3492] (#1946)
* fix

Fix for animators that have one layer and handle most animation state changes via properties.

* test and style

Adding an integration test to validate the changes.
Updated comments in NetworkAnimator.

* style

MTT-3492
Remove namespace references I was not using.

* style

only applying changes required for asmdef file.

* update and test

Removed the duplicate tracking for parameters.
Added the ability to make it owner or server authoritative.
Included triggers in this update.
Added test for triggers.

* style

* update

removing unused parameter state class

* update and fix

This updates the manual animation test so we can visually see NetworkAnimator working properly.
This fixes the issue where late joining clients would not be updated to the NetworkAnimator's state.
This fixes some test project asset related issues including to make sure we are always set to run in the background and have a target fps of 120.

* update and fix

This fixes the issue where it would sometimes not clean up/dispose the "cache" native array.
This improves the client synchronization (as best for the current way that we write information).

Manual Test Updates:
This fixes an issue with scale as well as getting other players to join and distribute evenly.
This fixes the issue with the prefab asset itself (scale issue impacted pulse).
Increased send/receive buffer sizes (we send a lot of data for animations currently)

* update

Making CleanUp private.

* fix

This lost its parameters for some reason...

* fix

Missed adding the trigger state back.

* test

This verifies that late joining clients get synchronized to the current NetworkAnimator state.
This also fixes the issue where stopping a client during an integration test always meant that you wanted to delete it.  Now, you have the option to not destroy in the event you want to stop a client, do some stuff, and then reconnect the client.

* test and style

Minor tweak and additional check that the AnimatorTestHelper instances equals the TotalClient count (more of a sanity check).
Added some comments and removed some LFs.

* Test and Style

I hooked up the TriggerTest to the wrong state.  Fixed issue by creating a new state that is activated by the test trigger.
Changing the warning to an actual error in TriggerTest (i.e. if that happens the test should fail).

* Update

This includes several adjustments to improve performance.
-Parameters are now sent separately
-New internal NetworkAnimatorStateChangeHandler that changes when:
  -It detects changes to the Animator
  -The server-side client forwarded RPC messages are processed here
-Removed the OwnerAuthoritative public bool

This also include integration test adjustments in order to be able to create and join a new client on the fly during a test.

* fix

This got test running locally in a stand alone testrunner build.
Going to give this a shot in Yamato land.

* wip

Sorting out the issue with server-side owner late joining client synchronization.  This version works with all tests ==but== the server-side owner late joining client portion... still trying to find this bug.

* fix

This reverts Triggers back to being set by an RPC (better to separate for now).
This also fixes minor issues with a player joining right in the middle of a transition.
This also fixes issue where trigger paths were not being set when transitioning (still could use more work).
This fixes the issue with the NetworkAnimator failing on the late joining player test where the late joining player could join while an Animator is in the middle of a transition.

* style

* style

* update

This improves upon the synchronization of a late joining client when connected while an animation is in the middle of a transition (it doesn't fix the issue completely, but is closer to being synchronized with the source animation)

* update

Removing VerboseDebug information

* Test

Added a small profiling test to profile where we spend time.
(sending RPCs)

* update - optimize

Optimizing how we send parameter updates.  Now we only send the parameters that changed.
This also avoids some smaller allocations by reusing a ClientRpcParams instance and a List<ulong> to hold the client ids to send to.

* style

updating a missed whitespace

* update

MTT-3492 MTT-2644
Updating what was fixed in the change log.

* update

Minor tweaks suggested by Fatih.