fix
This fixes an issue with dynamically spawned NetworkObjects throwing an exception if they are spawned during the awake method of the component spawning the objects due to the scene of origin handle still being 0 but the NetworkObject was spawned.  This should only throw an exception under those conditions if it is an in-scene placed NetworkObject.