fix
This is a new approach to synchronizing transitions with late joining players without having to actually set the associated conditional trigger.  By building a small list of all states and then building a quick lookup table, we can just synchronize the "transition state" which for late joining clients is a cross fade between the start and destination state.  We synchronize the normalized time (where it was in the transition on the server side when the client connected) of the transition this way as well.
Now, we just synchronize states (which some can be transition states).