Returned to a single NetworkVariable class. Removed all reflection and went back to an ILPP-based approach.
Limitation to this approach is that only NetworkVariables which are fields of NetworkBehaviours will get initialized... but as we don't really support any other usage of NetworkVariables, this seems like an acceptable limitation.

Also upgraded the project to 2020.3.40, as older versions fail to compile some of the new tests due to IL2CPP bugs.