fix
This is an experimental fix for the issue where you are rotating things in a nested hierarchy and you have changing parent-child scales.  The only way to properly synchronize rotation is by synchronizing the entire Quaternion.
This experiment provides a new NetworkTransform property, UseQuaternionSync, that will enable the synchronizing of the entire quaternion.
This still needs further thought on how we could potentially still provide a per axis synchronization.
This still needs to change how we synchronize rotation upon spawning a new NetworkObject on clients, so the NetworkTransform sends a full synchronization (currently) to all players when a new player joins (i.e. it is a work in progress).