fix
GetClientsWithStatus needed to add the host-client to the list of clients that completed once the scene event was completed.
Also fixed a bug in GetClientsWithStatus where, when checking for clients that did  not complete, it would include the clients that completed and the clients that did not complete in the list of client identifiers it returned.