test fix
The updates to SceneEventProgress exposed an issue with this integration test.
It was written before many NetcodeIntegrationTest helpers, so there are adjustments here to leverage from those which greatly simplifies the test.

Also, once this test and SceneEventProgress issues were resolved it exposed a potential QoL item where if a client does not validate a scene that contains in-scene placed NetworkObjects then it will cause a soft synchronization error on the client side.

I switched the scene that was not validated by the client-side to one that did not contain an in-scene placed NetworkObject to avoid this issue and to validate that a client can ignore loading a scene (as long as it does not have any in-scene placed NetworkObjects in it).