test
An updated version of this test that includes additional tests related to the refactoring.

This test was not properly testing the entire serialization pipeline and it only tested for one NetworkVariable.  If you added another NetworkVariable to the NetworkBehaviourWithNetworkVariables component the synchronization process would halt.  There was also a bug in SceneEventData that would occur if a NetworkObject failed to be instantiated. The updates included resolve all of these issues for the original intention behind the test.