fix
This fixes an issue with the NestedNetworkTransform test.  The children NetworkTransforms were still using world space synchronization which caused issues when the scale changed.  When that was corrected, it then exposed a bug in the delta position implementation where the predicted client position was not taking local space into consideration.
Also making sure k_QuaternionSync is not cleared after updating on the authoritative side.