test fix
Removing the sphere collider from the NetworkTransformOwnershipTests because the test basically ignores collision with other players and collision is not a factor of this test.  Since the spheres are effectively "doing nothing", it didn't make sense to try and add them when they weren't being used for anything and more importantly they were causing the client side to have skewed results when OwnerAuthoritativeTests was run under the HostStartsAsOwner pass.