update
Finally figured out the issue with rotating in a circle and interpolation.
We were always using Lerp which would interpolate in a linear (i.e. straight) line from axis value A to axis value B.
This is a temporary adjustment that provides one with the ability to determine if they want to be lerping or slerping their interpolated position.
(Really, we need to provide an interface to users that will allow them to build their own interpolators.)

This also includes an update during synchronization to better determine if it needs to update for HalfVector3 (delta position) or not...still not tested with owner authoritative.