update - tick replacing sent time
We now only send the network tick as opposed to the double precision server time since authority updates state only on each new network tick.  While there might be a very small factional difference between the actual time and the non-authoritative calculated time from the tick, the over-all end result is that it appears to be just as smooth when interpolating (if not slightly smoother).
Also removing the compression folder left out in the previous commit.