update - synchronize transform
This update allows users to determine if they want to synchronize the NetworkObject's associated transform when it is spawned.  This can help reduce the initial synchronization cost for things like in-scene placed NetworkObjects that have no visual components associated with them (i.e. managers etc) as well as "netcode aware" UI elements that have a pre-determined fixed position.