fix
Added tracking time out coroutines in order to stop them once a client is either approved or not but has not timed out.  Also only disconnecting a client that has timed out and but not disconnecting them if they were not approved (with the coroutine tracking it would be a very edge case scenario).