update - clean up
Cleaning up some of the NetworkManager start code to keep a more standard pattern.
Migrated the host-client specific code into its own method to help with this.
Updated the transport failure network event handler to call shutdown internal directly if  it occurs when starting the NetworkManager or just do the normal shutdown with stopping message processing if it occurs during a network session.