fix
This fixes an on-going issue with scale where really the answer was to send both the lossy and local scale values when synchronizing and/or teleporting and letting the client-side determine which to use depending upon whether the NetworkObject was parented, was the parenting applied when applying the states, and whether the parenting used WorldPositionStays.

This also fixes some minor issues with the parenting test not checking if the NetworkObject was parented before testing the final values.

This also updates named message handling slightly so it doesn't spam the log console during shutdown if there are incoming messages while shutting down.