feat(nested_network_objects): Added nested NetworkObject support
Introduces the concept of Dependent NetworkObjects. Dependent NetworkObjects cannot outlive the
NetworkObject they are dependent on. When spawning a prefab with nested NetworkObjects, all nested
NetworkObjects are dependent on the root NetworkObject. This means that nested NetworkObjects can
always be spawned/synchronized by spawning the root NetworkObject (and despawning any
NetworkObjects that are no longer spawned).

Resolves #2637