fix: GlobalObjectidHash serializing invalid values [MTT-77098] (#2662)
* fix
So far this user suggested fix looks promising. Creating a prefab from within a scene no longer requires one to delete the original (in-scene) object and then create a new instance from the newly created prefab.  It also is automatically detected and added to the default prefabs.
Fixing issue with namespace being outside of the UNIT_EDITOR conditional and removed an unused private property.
Fixing issue with PrefabStageUtility not existing in the same namespace based on Unity editor version.
Make sure the default network prefabs list is generated automatically by default.

* update
Minor rename of a variable.
Updated comments for further clarification as to what is happening.

* test
Adding method to expose GlobalObjectIdHash for manual testing purposes.
Attempting to fix the issue with TransformInterpolationTests stability.
Really, this test should be re-written.